refactor: move queries from other files to the queries folder
Closes #1169
