+++ backend/kernelCI_app/helpers/issueExtras.py
@@ -3,13 +3,13 @@ from typing import Dict, List, Set, Tuple
 
 from kernelCI_app.helpers.logger import log_message
 from kernelCI_app.helpers.trees import get_tree_url_to_name_map
+from kernelCI_app.queries.issues import get_issue_first_seen_data, get_issue_trees_data
 from kernelCI_app.typeModels.issues import (
     IssueWithExtraInfo,
     ProcessedExtraDetailedIssues,
     TreeSetItem,
     FirstIncident,
 )
-from kernelCI_app.models import Incidents, Checkouts
 
 
 class TagUrls:
@@ -48,46 +48,7 @@ def assign_issue_first_seen(
         issue_id_list.append(issue_id)
         versions_per_issue[issue_id].add(issue_version)
 
-    # TODO: use '=' instead of 'IN' when the list has a single element
-    incident_records = Incidents.objects.raw(
-        """
-        WITH first_incident AS (
-            SELECT
-                IC.issue_id,
-                MIN(IC.issue_version) AS min_issue_version,
-                MIN(IC._timestamp) AS first_seen
-            FROM
-                incidents IC
-            WHERE
-                IC.issue_id IN ({0})
-            GROUP BY
-                IC.issue_id
-        )
-        SELECT
-            IC.id,
-            IC.issue_id,
-            IC._timestamp AS first_seen,
-            IC.issue_version,
-            C.git_commit_hash,
-            C.git_repository_url,
-            C.git_repository_branch,
-            C.git_commit_name,
-            C.tree_name
-        FROM
-            incidents IC
-        LEFT JOIN tests T ON IC.test_id = T.id
-        LEFT JOIN builds B ON (IC.build_id = B.id OR T.build_id = B.id)
-        LEFT JOIN checkouts C ON B.checkout_id = C.id
-        JOIN first_incident FI ON (
-            IC.issue_id = FI.issue_id
-            AND IC.issue_version = FI.min_issue_version
-            AND IC."_timestamp" = FI.first_seen
-        )
-        """.format(
-            ", ".join(["%s"] * len(issue_id_list))
-        ),
-        issue_id_list,
-    )
+    incident_records = get_issue_first_seen_data(issue_id_list=issue_id_list)
 
     tree_url_to_name = get_tree_url_to_name_map()
     for record in incident_records:
@@ -127,53 +88,10 @@ def assign_issue_first_seen(
 
 def assign_issue_trees(
     *,
-    issue_key_list: List[Tuple[str, int]],
+    issue_key_list: list[tuple[str, int]],
     processed_issues_table: ProcessedExtraDetailedIssues,
 ) -> None:
-    # TODO: use '=' instead of 'IN' when the list has a single element
-    tuple_param_list = []
-    params = {}
-
-    for index, key in enumerate(issue_key_list):
-        id_key = f"id{index}"
-        version_key = f"version{index}"
-
-        tuple_string = f"(%({id_key})s, %({version_key})s)"
-
-        tuple_param_list.append(tuple_string)
-        params[id_key] = key[0]
-        params[version_key] = key[1]
-    tuple_str = ", ".join(tuple_param_list)
-
-    trees_records = Checkouts.objects.raw(
-        f"""
-        SELECT DISTINCT
-          ON (
-            C.TREE_NAME,
-            C.GIT_REPOSITORY_URL,
-            C.GIT_REPOSITORY_BRANCH,
-            IC.ISSUE_ID,
-            IC.ISSUE_VERSION
-          )
-          C.ID,
-          C.TREE_NAME,
-          C.GIT_REPOSITORY_URL,
-          C.GIT_REPOSITORY_BRANCH,
-          IC.ISSUE_ID,
-          IC.ISSUE_VERSION
-        FROM
-          INCIDENTS IC
-          LEFT JOIN TESTS T ON IC.TEST_ID = T.ID
-          LEFT JOIN BUILDS B ON (
-            IC.BUILD_ID = B.ID
-            OR T.BUILD_ID = B.ID
-          )
-          JOIN CHECKOUTS C ON B.CHECKOUT_ID = C.ID
-        WHERE
-          (IC.ISSUE_ID, IC.ISSUE_VERSION) IN ({tuple_str})
-        """,
-        params,
-    )
+    trees_records = get_issue_trees_data(issue_key_list=issue_key_list)
 
     tree_url_to_name = get_tree_url_to_name_map()
     for record in trees_records:
