+++ backend/kernelCI_app/helpers/trees.py
@@ -1,40 +1,14 @@
-from datetime import datetime
 import os
 
 from django.conf import settings
 import yaml
 from kernelCI_app.helpers.logger import log_message
-from kernelCI_app.models import Checkouts
 
 
 def make_tree_identifier_key(
     *, tree_name: str, git_repository_url: str, git_repository_branch: str
 ) -> str:
     return f"{tree_name}-{git_repository_url}-{git_repository_branch}"
-
-
-def get_tree_heads(origin: str, start_date: datetime, end_date: datetime):
-    tree_id_fields = [
-        "tree_name",
-        "git_repository_branch",
-        "git_repository_url",
-    ]
-
-    checkouts_subquery = (
-        Checkouts.objects.filter(
-            origin=origin,
-            start_time__gte=start_date,
-            start_time__lte=end_date,
-        )
-        .order_by(
-            *tree_id_fields,
-            "-start_time",
-        )
-        .distinct(*tree_id_fields)
-        .values_list("git_commit_hash", flat=True)
-    )
-
-    return checkouts_subquery
 
 
 def get_tree_url_to_name_map() -> dict[str, str]:
