+++ backend/kernelCI_app/queries/issues.py
@@ -4,7 +4,7 @@ from django.db.utils import ProgrammingError
 from kernelCI_app.helpers.database import dict_fetchall
 from kernelCI_app.helpers.environment import DEFAULT_SCHEMA_VERSION, set_schema_version
 from kernelCI_app.helpers.logger import log_message
-from kernelCI_app.models import Issues
+from kernelCI_app.models import Checkouts, Incidents, Issues
 from kernelCI_app.helpers.build import (
     is_valid_does_not_exist_exception,
     valid_status_field,
@@ -227,3 +227,106 @@ def get_test_issues(*, test_id: str) -> list[dict]:
         rows = dict_fetchall(cursor=cursor)
 
     return rows
+
+
+def get_issue_first_seen_data(*, issue_id_list: list[str]):
+    """Retrieves the incident and checkout data
+    of the first incident of a list of issues
+    through a list of `issue_id`s."""
+
+    # TODO: use '=' instead of 'IN' when the list has a single element
+    records = Incidents.objects.raw(
+        """
+        WITH first_incident AS (
+            SELECT
+                IC.issue_id,
+                MIN(IC.issue_version) AS min_issue_version,
+                MIN(IC._timestamp) AS first_seen
+            FROM
+                incidents IC
+            WHERE
+                IC.issue_id IN ({0})
+            GROUP BY
+                IC.issue_id
+        )
+        SELECT
+            IC.id,
+            IC.issue_id,
+            IC._timestamp AS first_seen,
+            IC.issue_version,
+            C.git_commit_hash,
+            C.git_repository_url,
+            C.git_repository_branch,
+            C.git_commit_name,
+            C.tree_name
+        FROM
+            incidents IC
+        LEFT JOIN tests T ON IC.test_id = T.id
+        LEFT JOIN builds B ON (IC.build_id = B.id OR T.build_id = B.id)
+        LEFT JOIN checkouts C ON B.checkout_id = C.id
+        JOIN first_incident FI ON (
+            IC.issue_id = FI.issue_id
+            AND IC.issue_version = FI.min_issue_version
+            AND IC."_timestamp" = FI.first_seen
+        )
+        """.format(
+            ", ".join(["%s"] * len(issue_id_list))
+        ),
+        issue_id_list,
+    )
+
+    return records
+
+
+def get_issue_trees_data(*, issue_key_list: list[tuple[str, int]]):
+    """Retrieves the list of trees in which a list of issues appear
+    through a list of tuples `issue_id, issue_version`.\n
+    The return is a single list where each element contains
+    the checkout data and the issue_id and issue_version of the incident."""
+
+    tuple_param_list = []
+    params = {}
+
+    for index, key in enumerate(issue_key_list):
+        id_key = f"id{index}"
+        version_key = f"version{index}"
+
+        tuple_string = f"(%({id_key})s, %({version_key})s)"
+
+        tuple_param_list.append(tuple_string)
+        params[id_key] = key[0]
+        params[version_key] = key[1]
+    tuple_str = ", ".join(tuple_param_list)
+
+    # TODO: use '=' instead of 'IN' when the list has a single element
+    records = Checkouts.objects.raw(
+        f"""
+        SELECT DISTINCT
+            ON (
+                C.TREE_NAME,
+                C.GIT_REPOSITORY_URL,
+                C.GIT_REPOSITORY_BRANCH,
+                IC.ISSUE_ID,
+                IC.ISSUE_VERSION
+            )
+            C.ID,
+            C.TREE_NAME,
+            C.GIT_REPOSITORY_URL,
+            C.GIT_REPOSITORY_BRANCH,
+            IC.ISSUE_ID,
+            IC.ISSUE_VERSION
+        FROM
+            INCIDENTS IC
+            LEFT JOIN TESTS T ON IC.TEST_ID = T.ID
+            LEFT JOIN BUILDS B ON (
+                IC.BUILD_ID = B.ID
+                OR T.BUILD_ID = B.ID
+            )
+            JOIN CHECKOUTS C ON B.CHECKOUT_ID = C.ID
+        WHERE
+            (IC.ISSUE_ID, IC.ISSUE_VERSION) IN ({tuple_str})
+        """,
+        params,
+    )
+
+    return records
